# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for EVE Ran application.",
        "title": "EVE Ran API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/characters": {
            "post": {
                "description": "Add a new character ID to the database and fetch all kills",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Add a new character ID",
                "parameters": [
                    {
                        "description": "Character ID",
                        "name": "character",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/characters/{id}": {
            "delete": {
                "description": "Remove a character from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Remove a character",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/characters/{id}/kills": {
            "get": {
                "description": "Fetch and store kills for a character from zKillboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Get character kills",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Kill"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/characters/{id}/kills/db": {
            "get": {
                "description": "Fetch kills for a character from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Get character kills from database",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/kills/region/{regionID}": {
            "get": {
                "description": "Fetch kills for a region from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kills"
                ],
                "summary": "Get kills by region",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Region ID",
                        "name": "regionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Attacker": {
            "type": "object",
            "properties": {
                "alliance_id": {
                    "type": "integer"
                },
                "character_id": {
                    "type": "integer"
                },
                "corporation_id": {
                    "type": "integer"
                },
                "damage_done": {
                    "type": "integer"
                },
                "faction_id": {
                    "type": "integer"
                },
                "final_blow": {
                    "type": "boolean"
                },
                "security_status": {
                    "type": "number"
                },
                "ship_type_id": {
                    "type": "integer"
                },
                "weapon_type_id": {
                    "type": "integer"
                }
            }
        },
        "models.Character": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Item": {
            "type": "object",
            "properties": {
                "flag": {
                    "type": "integer"
                },
                "item_type_id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Item"
                    }
                },
                "quantity_destroyed": {
                    "type": "integer"
                },
                "quantity_dropped": {
                    "type": "integer"
                },
                "singleton": {
                    "type": "integer"
                }
            }
        },
        "models.Kill": {
            "type": "object",
            "properties": {
                "attackers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Attacker"
                    }
                },
                "awox": {
                    "type": "boolean"
                },
                "character_id": {
                    "type": "integer"
                },
                "destroyedValue": {
                    "type": "number"
                },
                "droppedValue": {
                    "type": "number"
                },
                "fittedValue": {
                    "type": "number"
                },
                "hash": {
                    "type": "string"
                },
                "killmail_id": {
                    "type": "integer"
                },
                "killmail_time": {
                    "type": "string"
                },
                "locationID": {
                    "type": "integer"
                },
                "npc": {
                    "type": "boolean"
                },
                "points": {
                    "type": "integer"
                },
                "solar_system_id": {
                    "type": "integer"
                },
                "solo": {
                    "type": "boolean"
                },
                "totalValue": {
                    "type": "number"
                },
                "victim": {
                    "$ref": "#/definitions/models.Victim"
                }
            }
        },
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "models.Position": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                },
                "z": {
                    "type": "number"
                }
            }
        },
        "models.Victim": {
            "type": "object",
            "properties": {
                "alliance_id": {
                    "type": "integer"
                },
                "character_id": {
                    "type": "integer"
                },
                "corporation_id": {
                    "type": "integer"
                },
                "damage_taken": {
                    "type": "integer"
                },
                "faction_id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Item"
                    }
                },
                "position": {
                    "$ref": "#/definitions/models.Position"
                },
                "ship_type_id": {
                    "type": "integer"
                }
            }
        }
    }
}